+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 TestSwagger Changelog
 =====================
 
+### v0.1.4 (February 20, 2017)
+ - Renamed `SpyableObject` to `ObjectSpyable`
+ - Renamed `SpyableClass` to `ClassSpyable`
+
 ### v0.1.3 (February 17, 2017)
  - Added spy method forwarding behavior